+++ client/Packages/com.beamable/Runtime/BeamContext.cs
@@ -552,6 +552,7 @@ namespace Beamable
 				// _connectivityService.OnReconnectOnce( async () => await InitStep_GetUser());
 			}
 			AuthorizedUser.Value = user;
+			_requester.Language = user.language ?? _requester.Language;
 		}
 
 		private async Promise InitStep_StartPubnub()
@@ -595,8 +596,8 @@ namespace Beamable
 			// Create a new account
 			_requester.Token = _tokenStorage.LoadTokenForRealmImmediate(Cid, Pid);
 			_beamableApiRequester.Token = _requester.Token;
-			_requester.Language = SessionServiceHelper.GetISO639CountryCodeFromSystemLanague().ToLower();
-			
+			_requester.Language = SessionServiceHelper.GetISO639CountryCodeFromSystemLanguage().ToLower();
+
 			await InitStep_SaveToken();
 			await InitStep_GetUser();
 			var pubnub = InitStep_StartPubnub();